Updating the documentation registering the loader (#13)
* Updating the documentation registering the loader

The `registerLoader` function is deprecated, and it should be fixed in `Doctrine 2.0`. From the documentation it seems we should use it with the `class_exists`

From the documentation:
```
...
deprecated
this method is deprecated and will be removed in doctrine/annotations 2.0 autoloading should be deferred to the globally registered autoloader by then. For now, use @example AnnotationRegistry::registerLoader('class_exists')
```

* Update README.md